do not use UART TX FIFO
The ci20 UART TX FIFO is 64 bytes in size. In case the code crashes
and restarts for whatever reason, the actual debug printouts can lag
way behind, as the CPU is running much faster than the characters get
pushed out of the serial transmitter.

One way to solve this problem is to write characters to the TX FIFO
only when it is empty. This way the maximum output text loss is one
character.

Tested when debugging the remote reset command, observed all
expected characters printed before device reset.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>